Revert "Signup page"
