+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.glowskin_project">
    <application
-        android:label="GlowSkin"
+        android:label="glowskin_project"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
