--- lib/constans.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Palette {
-  static Color bg1 = Color(0xFFFFF9F1);
-  static Color bg2 = Color.fromRGBO(0, 0, 110, 1);
-}
