--- lib/users/favoritepage.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class FavoritePage extends StatefulWidget {
-  @override
-  State<FavoritePage> createState() => _FavoritePageState();
-}
-
-class _FavoritePageState extends State<FavoritePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-          child: Text(
-        'FavoritePage',
-      )),
-    );
-  }
-}
