--- lib/users/profilepage.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfilPage extends StatefulWidget {
-  @override
-  State<ProfilPage> createState() => _ProfilPageState();
-}
-
-class _ProfilPageState extends State<ProfilPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-          child: Text(
-        'ProfilPage',
-      )),
-    );
-  }
-}
