--- lib/users/signuppage.dart
@@ -1,131 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:glowskin_project/users/landingpage2.dart';
-import 'package:glowskin_project/users/launcher2.dart';
-
-class SignUpPage extends StatefulWidget {
-  const SignUpPage({Key? key}) : super(key: key);
-
-  @override
-  State<SignUpPage> createState() => _SignUpPageState();
-}
-
-class _SignUpPageState extends State<SignUpPage> {
-  TextEditingController nameController = TextEditingController();
-  TextEditingController emailController = TextEditingController();
-  TextEditingController passwordController = TextEditingController();
-  TextEditingController confirmPasswordController = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Padding(
-          padding: const EdgeInsets.all(10),
-          child: ListView(
-            children: <Widget>[
-              Container(
-                  alignment: Alignment.center,
-                  padding: const EdgeInsets.all(10),
-                  child: const Text(
-                    'GlowSkin',
-                    style: TextStyle(
-                        color: Colors.black,
-                        fontWeight: FontWeight.w500,
-                        fontSize: 25),
-                  )),
-              Container(
-                height: 300,
-                padding: const EdgeInsets.fromLTRB(80, 10, 0, 0),
-                margin: const EdgeInsets.fromLTRB(0, 20, 0, 20),
-                decoration: const BoxDecoration(
-                    image: DecorationImage(
-                  image: AssetImage(
-                    'assets/logo.png',
-                  ),
-                  fit: BoxFit.fitHeight,
-                )),
-              ),
-              Container(
-                padding: const EdgeInsets.all(10),
-                child: TextField(
-                  controller: emailController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(30)),
-                    ),
-                    labelText: 'User Name',
-                  ),
-                ),
-              ),
-              Container(
-                padding: const EdgeInsets.all(10),
-                child: TextField(
-                  controller: nameController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(30)),
-                    ),
-                    labelText: 'Email',
-                  ),
-                ),
-              ),
-              Container(
-                padding: const EdgeInsets.fromLTRB(10, 10, 10, 10),
-                child: TextField(
-                  obscureText: true,
-                  controller: passwordController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(30)),
-                    ),
-                    labelText: 'Password',
-                  ),
-                ),
-              ),
-              Container(
-                padding: const EdgeInsets.fromLTRB(10, 10, 10, 10),
-                child: TextField(
-                  obscureText: true,
-                  controller: confirmPasswordController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(30)),
-                    ),
-                    labelText: 'Confirm Password',
-                  ),
-                ),
-              ),
-              Container(
-                  height: 60,
-                  width: 125,
-                  padding: const EdgeInsets.fromLTRB(100, 10, 100, 0),
-                  child: ElevatedButton(
-                    child: const Text('Sign Up',
-                        style: TextStyle(
-                            fontSize: 20,
-                            color: Colors.white,
-                            fontWeight: FontWeight.w600)),
-                    style: ElevatedButton.styleFrom(
-                      primary: const Color(0xFF51AD99),
-                      onPrimary: Colors.white,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(30.0),
-                      ),
-                    ),
-                    onPressed: () {
-                      Navigator.of(context)
-                          .push(MaterialPageRoute(builder: (context) {
-                        return LauncherPage2();
-                      }));
-                      print(nameController.text);
-                      print(emailController.text);
-                      print(passwordController.text);
-                      print(confirmPasswordController);                      
-                    },
-                  )
-                  ),
-            ],
-          )
-          ),
-    );
-  }
-}
